+++ lib/src/core/core.dart
@@ -1,5 +1,6 @@
 export 'json_deserializer.dart';
 export 'json_formatter.dart';
+export 'json_info.dart';
 export 'json_serializer.dart';
 export 'json_type.dart';
 export 'xson_core.dart';